Move #number_of_friends to UesrDecorator from ProfileDecorator
