+++ app/decorators/user_decorator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module UserDecorator
+  def number_of_friends(friends)
+    friends.present? ? "(#{friends.size}人)" : ''
+  end
+end
