+++ app/views/profiles/index.html+mobile.erb
@@ -48,7 +48,7 @@
         </button>
       <% end %>
       <div>
-        <h3 class="text-2xl font-bold mb-2">知り合った人達<%= @profile.number_of_friends(@friends) %></h3>
+        <h3 class="text-2xl font-bold mb-2">知り合った人達<%= current_user.number_of_friends(@friends) %></h3>
         <ul class="mb-8">
           <% @friends.each do |friend| %>
             <li class="flex flex-row items-center">
